requirements.txt
python-decouple==3.6
Flask==2.2.2   
oauthlib==3.2.1
pyOpenSSL==22.1.0
Flask-Login==0.6.2
flask_mysqldb
